+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/collection_entries/QCertCreationAcceptance.kt
@@ -1,11 +1,15 @@
 package at.asitplus.rqes.collection_entries
 
 import at.asitplus.openid.TransactionData
+import at.asitplus.rqes.rdcJsonSerializer
+import at.asitplus.rqes.serializers.DeprecatedBase64URLTransactionDataSerializer
 import at.asitplus.signum.indispensable.asn1.ObjectIdSerializer
 import at.asitplus.signum.indispensable.asn1.ObjectIdentifier
 import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.JsonPrimitive
 
 /**
  * D3.1: UC Specification WP3:
@@ -77,6 +81,12 @@ data class QCertCreationAcceptance(
 
 ) : TransactionData {
 
+    override fun toBase64UrlString(): JsonPrimitive =
+        rdcJsonSerializer.parseToJsonElement(
+            rdcJsonSerializer.encodeToString(
+                DeprecatedBase64URLTransactionDataSerializer, this
+            )
+        ) as JsonPrimitive
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true