+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/PresentationDataClasses.kt
@@ -5,12 +5,11 @@ import at.asitplus.dif.ConstraintField
 import at.asitplus.dif.PresentationSubmission
 import at.asitplus.jsonpath.core.NodeList
 import at.asitplus.jsonpath.core.NormalizedJsonPath
-import at.asitplus.openid.TransactionData
+import at.asitplus.openid.TransactionDataBase64Url
 import at.asitplus.openid.dcql.DCQLCredentialQueryIdentifier
 import at.asitplus.signum.indispensable.cosef.CoseSigned
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import at.asitplus.signum.indispensable.josef.JwsSigned
-import at.asitplus.wallet.lib.data.DeprecatedBase64URLTransactionDataSerializer
 import at.asitplus.wallet.lib.data.VerifiablePresentationJws
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.iso.sha256
@@ -34,7 +33,7 @@ import kotlinx.serialization.json.buildJsonObject
 data class PresentationRequestParameters(
     val nonce: String,
     val audience: String,
-    val transactionData: Pair<Flow, Collection<TransactionData>>? = null,
+    val transactionData: Pair<Flow, List<TransactionDataBase64Url>>? = null,
     /**
      * Handle calculating device signature for ISO mDocs, as this depends on the transport protocol
      * (OpenId4VP with ISO/IEC 18013-7)
@@ -49,15 +48,6 @@ data class PresentationRequestParameters(
         OID4VP,
         UC5
     }
-
-    internal fun getTransactionDataHashes(): Set<ByteArray>? = transactionData?.second?.map {
-        (vckJsonSerializer.encodeToJsonElement(
-            DeprecatedBase64URLTransactionDataSerializer,
-            it
-        ) as JsonPrimitive).content.decodeToByteArray(
-            Base64UrlStrict
-        ).sha256()
-    }?.toSet()
 }
 
 sealed interface PresentationResponseParameters {